scene transfer cameras fixed
gave camera to second room. 'room' prefab has camera w room polygon.

you still need a main camera w cinamachine brain on it, and audio listener on it too.

if it breaks when transitioning scenes. make sure all 'rooms' are 'turned on' (their name isn't greyed out), but make sure the secondary vcams ARE 'turned off' (the name will be greyed out). the cinemachine is handling those transitions and turning stuff off and on for you.