singleton for doors
doors can save a bool for if they are open now. bug fixed with sorting error for the Received Item, now displays in front.

Bug fixed for keybox (kind of) by getting rid of the text item that was messing up the unlocking deal. There is no text for the mushroom item now but its not important