pause menu
the pause menu works (with clicking). also the 'pause panel' has to be turned off when you start the game, in the inspector.

i aspire to make things work on the space bar or on enter at some point. i deleted the submit function before, i have remade it now but it had the knock-on effect of making space bar for enter not work.