Fixed sitting fixed walking
all the states have their 'stop moving' parts fixed. the references to speed = 0; were breaking it because speed was never assigned fast again.

Instead we made 'change' 0;, which does change when we go into move state.

also to 're-set' the bools on the boxes. you need to enter playmode, tick or untick those bools in the inspector, exist, then re-enter playmode. that is just how it is.

speed was changed to 4, the uncanny 'gliding' effect was caused by the speed being too high. it is literally perfect.