Extension of native class error fixed
Having 'savedata' as the name of the script, and the name of the function was making it angry.
I had applied Schwer's script to a gameobject. it does not need to be applied to gameobject. it was doing its thing in the namespace.

do not apply namespace things randomly to objects, they are happy by themselves.