keyDoor script updated
now works with schwer system. tested on keyBox. need to make actual keydoor that animates similarly and test it on that.

appears to work with Specific key which I prefer to Taft/zelda-like system.

item description text does not display when you open a keybox, but it was already like that. that's just a mystery.