sprite changer improved
sprites can be assigned randomly out of a list, only once, on start.

I mean, you could get it to happen randomly every however many seconds if you wanted, I guess you would use a coroutine for that though.

You can also set a specific sprite. At the moment, this gets set in the update method.  But my idea is that something outside of this script could call it.

Eventually I was thinking of an enum for like, emotion states, and something that sets it externally, for NPC for example.