prefab of emotional object added
called 'emotion mushroom'. To use: press space, it will change its 'emotion' to sad, and the sprite with change animation accordingly.

possible extensions: add schwer namespace at the top, and make if the player has such and such an inventory in the possession, then a certain emotion would show.

you would have to call it sadItem in the inspector, like set a public item called sadItem,  and check in the fixed update if they had it in their player inventory.

but all of that would be better if it only updated when it was 'interacted' with, like exending the interact class. it cant be healthy if the computer is checking it at fixed updates.

actually i guess having the bool for 'sprite changed' helped with that, you can use it to mean that the sprite is assigned once initially, like start on idle and only check anything if it's not on idle