Prior required conversation
conversation item can be attached that requires a different conversation to have been completed in the past.

have not tested: two conversations placed on one NPC, with one dependent on the other. If this is an issue, a coroutine taking place for dependent conversations might be necessary.

Mild issues with getting the 'heart' script on DropItem.cs items - how can I make it usable, when using the item forces me to point to a game object for a script (eg health up script). When we pick up an item, it gets destroyed.

(It seems like a destroyed item, i wouldnt be able to use the code on it, since it's destroyed.)