conversation error fixed
'third convo', the scriptable object for the conversation, was in the game save manager. it needed the 'third convo COMPLETED' bool to go in there instead.

basically, it was treating the conversation SO like it was a bool, so that was wiping all the info that the speaker UI needed on runtime.