Moved code over to DialogueNPC
error is 'Assets/cwg/scripts/cwgpm scripts/ScriptableObjects/Dialogs/DialogNPC.cs(66,104): error CS0120: An object reference is required for the non-static field, method, or property 'DialogNPC.requiredCompleted'
'

slight change from 'randomscript'. The BoolValue is not the thing that decides the dropdown inspector menu. It lives inside the dropdown menu and contains the conversation that we have to have finished before the dependent one triggers.