gamesaver prefabs updated
in the *SCENES THEMSELVES* gamesaver has the right stuff in now. (there will be some demo/blank/test scene that does not have this, needs to be deleted).

You (remy) should probably go into the prefab object itself for gamesaver and make sure it has the right stuff in. The cute notebook in real life has the dependencies for this written down.