UI canvas scale fixed
"First, make sure the CanvasScaler component on the Canvas has the "UI Scale Mode" set to "Scale with Screen Size", "Screen Match Mode" set to "Match Width or Height" and the Match slider set all the way to the smallest axis (if the resolutions is set to 800x600, set the slider to 1). This will make the canvas to be almost always as big as the whole screen, respecting the aspect ratio.

Next, select the Panel, look at the Image component. Set the Source Image to "none". This will paint the canvas as a full rectangle of one color. Now in the Rect Transform set the anchor mode to stretch in both directions, and set Left, Right, Bottom and Top to 0. Reset the scale to 1,1,1.

This should do it."