achievement panel displays correctly
the bool value 'achieve1' tracks whether or not you have unlocked the achievement in the crafting panel.

If you have already seen it, it won't come up again if you leave and re-enter the scene.

Image and text needed to be children of an empty gameobject which holds the script: a script can't turn a gameobject on if it is attached to that same gameobject which is already 'off'.

That's why the gameobjects are children.