dog visitor disappears after one appearance
After you talk to him once, if you come back in the restaurant, he's gone (this is what I want).
Achieved this by putting dog visitor disappearer on the dog visitor himself BUT ALSO the timeline manager. both make him appear.

I originally wanted him to walk out after you finish a conversation, but the bool for 'convo completed' marks them as completed when they begin, not when they end. this is a little irritating.